refactor/디렉토리 구조 refactor(pages/PostCommunity, components/PostCommunity)
