+++ frontend/src/pages/PostCommunity/MyShots.vue
